Rename `download` to `fetch`, reduce scope of fetching
In reality, the only thing that needs to be implemented when extending
`Scraper[F[_]]` is the HTTP fetching logic. Making such method only
fetch the main client HTML is too restrictive.

This way, we can also reuse the `fetch` method to fetch the required
script resource in order to get the build number.