Make http4s Client[F] the resource to pass around
A Scraper[F] doesn't really require allocation and teardown -- it's just
a handful of functions that operate on Asset objects and strings. The
actual object that needs to function as a resource is the http4s
Client[F].

What prompted this is the incoming addition of publishers, which also
have to make HTTP requests. However, the HTTP requesting is only done in
Scraper[F]. Let's lift it into a Client[F] resource so that the
publisher can use it as well.