Use MonadError instead of EitherT
MonadError[F[_], E] is more composable than EitherT[F[_], A, B], so
let's use that instead. That way, we can also eliminate most of the
leftMap calls and other error plumbing that we needed to do.

This patch also refactors the arrangement of Main once more and deletes
custom exceptions that are no longer needed.