Extract "Version Hash" from builds
This functionality was added into fetchBuildNumber then renamed into
fetchBuildInfo, since the regex used already captures the "Version
Hash".

NoBuildNumber was also renamed to NoBuildInfo. I was considering making
a BuildInfo case class containing the hash and the build number, but I'm
not so sure about that right now. Maybe later. If we do, we can make
fetchBuildInfo return an instance of it instead of a tuple like this
patch does.