Persist last known builds to a state file
Write a state file to disk every time a new build is published, so we
don't announce that build once the program is restarted for whatever
reason.

The state file holds a mapping of source selector to last known build
number in a simple key-value format. Internally, this mapping is
referred to as the "state store".

With the necessary groundwork in place, all it took was an additional
`.scan` call in the main polling stream, alongside piping it to the
file.

In order for this to be done, we need to establish a two-way
relationship between sources and source selectors. The `Select` type
class had to be changed to hold a `Map` instead of a `Set`. This is so
that if the user gave a wildcard (also referred to as a "star glob"), we
are able to derive the selectors. (Significant because we save them into
the state file.) In the previous design, a method that would resolve
items in a `Set` from a string didn't allow other code to access those
raw string keys, while a `Map` does.

Finally, because `Source` doesn't track the selector string that
selected it, a `SelectedSource` case class was introduced in order to
track both at the same time. (Again, significant because we end up
saving the selector into the state file.)

Closes #7.