Add host build source
Add a source for scraping host builds. A host is the actual Electron
desktop application.

In order to support multiple types of builds, Build was converted into
an ADT, and the B type parameter in Source[F[_], B] was made covariant.
Also, the comparison type class for builds was switched back to Eq from
Order.

Closes #4.