Fix "last-known deduplication" only working once
While writing tests, I realized that the first build *is* deduplicated
if its version matches the one saved in the state file on disk, but it's
only deduplicated *once* instead of every time that build with that
version is emitted.

This is because the deduplication is happening before the stream gets
`.changes` called on it, which happens in `poll`. We need to move the
deduplication *after* the `poll` call. Do that, and also clean up and
refactor the surrounding code to make things pleasant and nice.