Improve selector-related errors and nomenclature
Let's start using the terms "variant" and "source" when it comes to
selector parsing, as it unambiguously refers to specific parts of the
selector. The rest of the codebase should get this treatment soon.

Now we have the proper terminology to introduce good error handling, so
do that.

An ADT for every possible source was also introduced. This should be
moved into the proper package as soon as possible.